+++ .php-cs-fixer.dist.php
@@ -1,24 +1,28 @@
 <?php
 /*
- * This file is part of Phyxo package
+ *  -- BEGIN LICENSE BLOCK ----------------------------------
  *
- * Copyright(c) Nicolas Roudaire  https://www.nikrou.net/
- * Licensed under the GPL version 2.0 license.
+ *  This file is part of externalLinks, a plugin for DotClear2.
  *
- * For the full copyright and license information, please view the COPYING
- * file that was distributed with this source code.
+ *  Licensed under the GPL version 2.0 license.
+ *  See COPYING file or
+ *  http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
+ *
+ *  -- END LICENSE BLOCK ------------------------------------
  */
 
 // the final goal is to respect all PSR-2 rules.
 
 $header = <<<'EOF'
-This file is part of Phyxo package
+ -- BEGIN LICENSE BLOCK ----------------------------------
 
-Copyright(c) Nicolas Roudaire  https://www.nikrou.net/
-Licensed under the GPL version 2.0 license.
+ This file is part of externalLinks, a plugin for DotClear2.
 
-For the full copyright and license information, please view the COPYING
-file that was distributed with this source code.
+ Licensed under the GPL version 2.0 license.
+ See COPYING file or
+ http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
+
+ -- END LICENSE BLOCK ------------------------------------
 EOF;
 
 $finder = PhpCsFixer\Finder::create()->in(__DIR__);